Move miscFunctions
