+++ src/js/chat.js
@@ -48,14 +48,6 @@ if ((document.URL.indexOf("https://habitica.com/apidoc/") == -1) && (document.UR
   };
   (document.head||document.documentElement).appendChild(s);
 
-  // Call other functions
-  var s = document.createElement('script');
-  s.src = browser.getExtensionURL('resources/miscFunctions.js');
-  s.onload = function() {
-      this.parentNode.removeChild(this);
-  };
-  (document.head||document.documentElement).appendChild(s);
-
   // Call main chat script
   //pause 1 seconds to allow everything to catch up
   setTimeout(function() {
