+++ src/js/lib/browser.js
@@ -34,5 +34,6 @@ function createTab(url) {
 module.exports = {
   getStorage,
   setStorage,
-  getExtensionURL
+  getExtensionURL,
+  createTab
 };
